refactor: 패키기 구조 변경 및 삭제
