--- src/main/java/com/trashheroesbe/feature/gpt/dto/request/TrashAnalysisRequestDt.java
@@ -1,6 +1,0 @@
-package com.trashheroesbe.feature.gpt.dto.request;
-
-public record TrashAnalysisRequestDt(
-        String imageUrl
-) {
-}
